+++ codegen/json/json-schema/src/main/groovy/com/stanfy/helium/handler/codegen/json/schema/JsonType.java
@@ -7,7 +7,7 @@ import java.util.Locale;
  *
  * @author Michael Pustovit mpustovit@stanfy.com.ua
  */
-enum JsonType {
+public enum JsonType {
   ARRAY,
   BOOLEAN,
   INTEGER,