+++ codegen/swagger/src/main/java/com/stanfy/helium/swagger/Definition.java
@@ -1,0 +1,5 @@
+package com.stanfy.helium.swagger;
+
+/** Swagger definition. */
+final class Definition {
+}