+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.8.1
+-----
+*(not released yet)*
+
+- Basic Swagger integration
+- Maps support (dictionaries) in helium spec
+- Improved validation messages
+
 0.6.0
 -----
 *(2015-05-29)*